// lib/generated/inputTypeSchemas/AmenitiesUpdateInputSchema.ts
import type { Prisma } from '../../prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { AMENITIESTYPESchema } from './AMENITIESTYPESchema';
import { EnumAMENITIESTYPEFieldUpdateOperationsInputSchema } from './EnumAMENITIESTYPEFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { HotelUpdateOneWithoutAmenitiesNestedInputSchema } from './HotelUpdateOneWithoutAmenitiesNestedInputSchema';

export const AmenitiesUpdateInputSchema: z.ZodType<Prisma.AmenitiesUpdateInput> = z.object({
  id: z.union([ z.string().uuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  description: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  icon: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  image: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  amenitiestype: z.union([ z.lazy(() => AMENITIESTYPESchema),z.lazy(() => EnumAMENITIESTYPEFieldUpdateOperationsInputSchema) ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  Hotel: z.lazy(() => HotelUpdateOneWithoutAmenitiesNestedInputSchema).optional()
}).strict();

export default AmenitiesUpdateInputSchema;